@@ -4,12 +4,12 @@ version: 1.0.0+1
 homepage: null
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
 
 dependencies:
-  http: ^0.13.4
+  http: ^1.2.2
 
 dev_dependencies:
   coverage: ^1.0.3
-  mocktail: ^0.2.0
+  mocktail: ^1.0.4
   test: ^1.16.4
