@@ -4,11 +4,11 @@ version: 1.0.0+1
 publish_to: none
 
 environment:
-  sdk: ">=2.15.1 <3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
 
 dependencies:
-  freezed_annotation: ^1.1.0
-  http: ^0.13.4
+  freezed_annotation: ^2.4.4
+  http: ^1.2.2
   json_annotation: ^4.4.0
   base_api:
     path: ../base_api
@@ -16,7 +16,7 @@ dependencies:
 dev_dependencies:
   build_runner: ^2.1.7
   coverage: ^1.0.3
-  freezed: ^1.1.0
+  freezed: ^2.5.7
   json_serializable: ^6.1.5
-  mocktail: ^0.2.0
+  mocktail: ^1.0.4
   test: ^1.20.1
