@@ -3,10 +3,9 @@ import 'package:chingu_bookfinder_flutter/book/book.dart'
     show BookDetailBloc, BookListBloc, BookListState, GetBookDetailEvent;
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:go_router/go_router.dart';
 
 class BookList extends StatelessWidget {
-  const BookList({Key? key}) : super(key: key);
+  const BookList({super.key});
 
   @override
   Widget build(BuildContext context) {
